+++ extensions/core/interfaces/wysiwyg-full/readonly.vue
