+++ extensions/core/layouts/tabular/meta.json
