+++ public/extensions/.gitignore
@@ -1,3 +1,3 @@
 core/interfaces
-core/listings
+core/layouts
 core/pages
