+++ src/core/Directus/Services/LayoutsService.php
@@ -4,14 +4,14 @@ namespace Directus\Services;
 
 use Directus\Application\Container;
 
-class ListingsService extends AbstractExtensionsController
+class LayoutsService extends AbstractExtensionsController
 {
     public function __construct(Container $container)
     {
         parent::__construct($container);
 
         $basePath = $this->container->get('path_base');
-        $this->basePath = $basePath . '/public/extensions/core/listings';
+        $this->basePath = $basePath . '/public/extensions/core/layouts';
     }
 
     public function findAll(array $params = [])
